refactor: navigation file structure
