+++ web/src/components/Navigation/NavIconButton/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './NavIconButton';
