+++ web/src/components/Navigation/NavLink/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './NavLink';
