+++ web/src/components/Navigation/SecondaryNavigations/index.tsx
@@ -1,0 +1,3 @@
+export { default as ComplianceNavigation } from './ComplianceNavigation';
+export { default as LogAnalysisNavigation } from './LogAnalysisNavigation';
+export { default as SettingsNavigation } from './SettingsNavigation';
